requirements.txt
requests
pendulum
psycopg2-binary
apache-airflow-providers-postgres